requirements.txt
arcade==1.0.1.post7
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
EasyProcess==0.2.3
flake8==3.5.0
future==0.16.0
idna==2.6
lxml==4.1.1
mccabe==0.6.1
numpy==1.12.1
olefile==0.44
pew==1.1.2
Pillow==4.0.0
pipenv==9.0.3
psutil==5.3.1
pycodestyle==2.3.1
pyflakes==1.6.0
pyglet==1.2.4
python-telegram-bot==9.0.0
PyVirtualDisplay==0.2.1
requests==2.18.4
selenium==3.8.1
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
